+++ rmk-config/CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Update toml config fields
-- direct_pin_matrix section
+- Add direct pin matrix configs
 
 ## [0.1.6] - 2024-10-03